+++ html-formatter/java/src/test/java/io/cucumber/htmlformatter/Jackson.java
@@ -1,9 +1,10 @@
-package io.cucumber.messages;
+package io.cucumber.htmlformatter;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.cfg.ConstructorDetector;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 
@@ -11,11 +12,12 @@ import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKN
 import static com.fasterxml.jackson.databind.DeserializationFeature.READ_ENUMS_USING_TO_STRING;
 import static com.fasterxml.jackson.databind.SerializationFeature.WRITE_ENUMS_USING_TO_STRING;
 
-public final class Jackson {
+final class Jackson {
     public static final ObjectMapper OBJECT_MAPPER = new ObjectMapper()
             .registerModule(new ParameterNamesModule(JsonCreator.Mode.PROPERTIES))
             .registerModule(new Jdk8Module())
             .setSerializationInclusion(JsonInclude.Include.NON_ABSENT)
+            .setConstructorDetector(ConstructorDetector.USE_PROPERTIES_BASED)
             .enable(WRITE_ENUMS_USING_TO_STRING)
             .enable(READ_ENUMS_USING_TO_STRING)
             .disable(FAIL_ON_UNKNOWN_PROPERTIES)
@@ -24,3 +26,4 @@ public final class Jackson {
     private Jackson() {
     }
 }
+