+++ messages/java/src/main/java/io/cucumber/messages/MessageWriter.java
@@ -2,6 +2,6 @@ package io.cucumber.messages;
 
 import java.io.IOException;
 
-public interface MessageWriter {
-    void write(Object message) throws IOException;
+public interface MessageWriter<T> {
+    void write(T message) throws IOException;
 }